Change Float32 to Float 64

+++ src/Models/topological_elements.jl
@@ -3,9 +3,9 @@ struct Junction
     number::Int
     name::String
     elevation::Real
-    head::Union{Nothing,Float32}
+    head::Union{Nothing,Float64}
     demand::Any
-    minimum_pressure::Float32
+    minimum_pressure::Float64
     coordinates::Union{@NT(lat::Float64, lon::Float64),Nothing}
 end
 
@@ -14,7 +14,7 @@ Junction(;  number = 0,
             elevation=0,
             head = nothing,
             demand = nothing,
-            minimum_pressure = 0,
+            minimum_pressure = 0.0,
             coordinates = nothing
             ) = Junction(number,name,elevation,head,demand,minimum_pressure,coordinates)
